Added documentation on datatypes and exporters, and fixed a few bugs:
- cmdline arguments specific to 'band export' are now shown
  only for band exports, and not for all exports on the verdi data
  cmdline
- Prettifier class moved from Bands to Kpoints
- Moved the encoding to the single _prepare_ routines; this is needed
  because e.g. when exporting a PDF, you cannot encode it using utf-8
- made get_export_formats a classmethod
- Added long documentation on existing data classes, and export functionality
