+++ docs/source/developer_guide/devel_tutorial/code_plugin_float_sum.rst
@@ -1,3 +1,5 @@
+.. _DevelopDataPluginTutorialFloat:
+
 Developer data plugin tutorial  - Float summation
 =================================================
